순한 맛 1 번끝.
2번 끝

3.주문하기

+++ src/main/java/com/bigtyno/restaurant/dto/FoodDto.java
@@ -9,5 +9,5 @@ import lombok.NoArgsConstructor;
 @Getter
 public class FoodDto {
     private String name;
-    private Long price;
+    private int price;
 }
